+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Additions and changes from [pull request #289](https://github.com/codemagic-ci-c
 **Features**
 
 - Add new action `app-store-connect review-submissions list` to list existing review submissions in the App Store. See official API method [documentation](https://developer.apple.com/documentation/appstoreconnectapi/get_v1_reviewsubmissions).
-- Add new action `app-store-connect review-submissions cancel-review-submissions` to cancel existing review submissions in the App Store based on their type. Uses the already existing `app-store-connect review-submissions cancel` action internally that allows to set the submission status to `canceled` using `PATCH`. See official API method [documentation](https://developer.apple.com/documentation/appstoreconnectapi/patch_v1_reviewsubmissions_id).
+- Add new action `app-store-connect app cancel-review-submissions` to cancel existing review submissions in the App Store based on their type for a specific app. Uses the already existing `app-store-connect review-submissions cancel` action internally that allows to set the submission status to `canceled` using `PATCH`. See official API method [documentation](https://developer.apple.com/documentation/appstoreconnectapi/patch_v1_reviewsubmissions_id).
 - Add new action `app-store-connect builds expire` to expire a specific build that has been uploaded to App Store Connect. Modifies the existing build resource to an expired status using `PATCH`. See official API method [documentation](https://developer.apple.com/documentation/appstoreconnectapi/modify_a_build).
 - Add new action `app-store-connect apps expire-builds` to expire all builds uploaded to App Store Connect except the given build(s) for the specific application. Uses the aforementioned `app-store-connect builds expire` action internally.
 - Add new action `app-store-connect builds expire-build-submitted-for-review` to expire build in App Store Connect that has been submitted to review and has not been `Approved`. Uses the aforementioned `app-store-connect builds expire` action internally.
