+++ docs/app-store-connect/apps/cancel-review-submissions.md
@@ -1,12 +1,12 @@
 
-review-submissions
-==================
+cancel-review-submissions
+=========================
 
 
-**Manage your App Store version review submissions**
+**Find and cancel review submissions in App Store Connect for the given application**
 ### Usage
 ```bash
-app-store-connect review-submissions [-h] [--log-stream STREAM] [--no-color] [--version] [-s] [-v]
+app-store-connect apps cancel-review-submissions [-h] [--log-stream STREAM] [--no-color] [--version] [-s] [-v]
     [--log-api-calls]
     [--api-unauthorized-retries UNAUTHORIZED_REQUEST_RETRIES]
     [--api-server-error-retries SERVER_ERROR_RETRIES]
@@ -17,8 +17,26 @@ app-store-connect review-submissions [-h] [--log-stream STREAM] [--no-color] [--
     [--private-key PRIVATE_KEY]
     [--certificates-dir CERTIFICATES_DIRECTORY]
     [--profiles-dir PROFILES_DIRECTORY]
-    ACTION
+    [--platform PLATFORM]
+    [--review-submission-state REVIEW_SUBMISSION_STATE]
+    APPLICATION_ID_RESOURCE_ID
 ```
+### Required arguments for action `cancel-review-submissions`
+
+##### `APPLICATION_ID_RESOURCE_ID`
+
+
+Application Apple ID. An automatically generated ID assigned to your app
+### Optional arguments for action `cancel-review-submissions`
+
+##### `--platform, --app-store-version-platform=IOS | MAC_OS | TV_OS`
+
+
+App Store Version platform. Default:&nbsp;`IOS`
+##### `--review-submission-state=CANCELING | COMPLETE | COMPLETING | IN_REVIEW | READY_FOR_REVIEW | UNRESOLVED_ISSUES | WAITING_FOR_REVIEW`
+
+
+String value of the review submission state. Multiple arguments
 ### Optional arguments for command `app-store-connect`
 
 ##### `--log-api-calls`
@@ -87,12 +105,3 @@ Disable log output for commands
 
 
 Enable verbose logging for commands
-### Actions
-
-|Action|Description|
-| :--- | :--- |
-|[`cancel`](review-submissions/cancel.md)|Discard a specific review submission from App Review|
-|[`confirm`](review-submissions/confirm.md)|Confirm pending review submission for App Review|
-|[`create`](review-submissions/create.md)|Create a review submission request for application's latest App Store Version|
-|[`get`](review-submissions/get.md)|Read Review Submission information|
-|[`list`](review-submissions/list.md)|Find and list review submissions in App Store Connect for the given application|
