+++ src/codemagic/tools/_app_store_connect/action_groups/review_submissions_actions_group.py
@@ -82,33 +82,6 @@ class ReviewSubmissionsActionGroup(AbstractBaseAction, metaclass=ABCMeta):
         )
 
     @cli.action(
-        'cancel-review-submissions',
-        AppArgument.APPLICATION_ID_RESOURCE_ID,
-        AppStoreVersionArgument.PLATFORM,
-        ReviewSubmissionArgument.REVIEW_SUBMISSION_STATE,
-        action_group=AppStoreConnectActionGroup.REVIEW_SUBMISSIONS,
-    )
-    def cancel_review_submissions(
-        self,
-        application_id: ResourceId,
-        platform: Optional[Platform] = None,
-        review_submission_state: Optional[Union[ReviewSubmissionState, Sequence[ReviewSubmissionState]]] = None,
-        should_print: bool = False,
-    ) -> List[ReviewSubmission]:
-        """
-        Find and cancel review submissions in App Store Connect for the given application
-        """
-
-        review_submissions = self.list_review_submissions(
-            application_id,
-            platform,
-            review_submission_state,
-            should_print,
-        )
-
-        return [self.cancel_review_submission(submission.id) for submission in review_submissions]
-
-    @cli.action(
         'confirm',
         ReviewSubmissionArgument.REVIEW_SUBMISSION_ID,
         action_group=AppStoreConnectActionGroup.REVIEW_SUBMISSIONS,
